Extract 'model' to a dedicated module
