+++ solution/solution.cabal
@@ -21,10 +21,13 @@ executable solution
     , warp                 >= 3.2.28 && < 3.3
 
   ghc-options:
-    -Wall
+    -Wall -O -threaded -rtsopts -with-rtsopts=-N
 
   default-language:
     Haskell2010
 
   main-is:
     Main.hs
+
+  other-modules:
+    Model
